requirements.txt
langchain==0.1.0
langchain-openai==0.0.5
langgraph==0.0.20
weaviate-client==4.4.0
python-dotenv==1.0.0
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.0
httpx==0.26.0
phonenumbers==8.13.27
langdetect==1.0.9
twilio==8.0.0
openai>=1.0.0
